Move TicTacToe engine to the environment directory
