+++ training_platform/clients/basic_player_clients/rl_player_client.py
@@ -14,7 +14,7 @@ from reinforcement_learning.agents.basic_agent import BasicAgent
 from training_platform.server.common import *
 from training_platform.server.service import GameManager, MatchMaker
 from training_platform.server.logger import Logger
-from game_app.games.tic_tac_toe.engine.tic_tac_toe_engine_utils import Player
+from environments.tic_tac_toe.tic_tac_toe_engine_utils import Player
 
 def signal_handler(sig, frame):
     asys.tell(match_maker_addr, DetachMsg())
